chore: Accomodate `mock_prover_verify` to not return
Since the new release of halo2_2023_10_17, `VerifyFailure` now has a
reference to the prover itself.

That means that we can't return the failures if we are out of the prover
scope.
To adapt to the situation, the fn `mock_prover_verify` has been changed
so that does the assertion inside.

Resolves: #60