Add license to match other pse projects
This repository is currently missing any licensing information.  Given
that this is a privacy-scaling-explorations project and the others seem
to have adopted a dual Apache/MIT licensing approach, I've copied those
pieces here.

Please advise if the licensing was intentionally left blank for some
reason, but it seems like a simple oversight to me.